feat: Phase 1.3 FSRS 集成完成
## FSRS-4.5 算法实现（纯 Rust）

### FsrsEngine
- 基于 FSRS-4.5 算法（无外部依赖）
- 17个权重参数（默认值）
- 完整的稳定性/难度计算

### 核心功能
- schedule_review() - 计算下次复习
- preview_ratings() - 预览4种评分结果
- forgetting_curve() - 遗忘曲线
- should_review() - 是否应该复习
- overdue_days() - 逾期天数

### 评分系统
- Again: 完全不记得 → 重置间隔
- Hard: 困难 → 小幅增加
- Good: 良好 → 正常增加
- Easy: 简单 → 大幅增加

### 状态跟踪
- stability (稳定性)
- difficulty (难度 1-10)
- reps (复习次数)
- lapses (遗忘次数)
- next_review (下次复习时间)

### 修复
- CardState: f32 → f64 (精度提升)
- Rating: 添加 PartialEq, Eq

### 测试
- 7个单元测试覆盖
- 初始状态/首次复习/遗忘/预览/难度

编译通过 ✅
无外部 crate 依赖（避免冲突）

+++ src-tauri/src/domain/event.rs
@@ -222,7 +222,7 @@ pub struct Annotation {
 }
 
 /// FSRS 评分
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
 pub enum Rating {
     Again, // 完全不记得
@@ -234,8 +234,8 @@ pub enum Rating {
 /// FSRS 卡片状态
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct CardState {
-    pub stability: f32,
-    pub difficulty: f32,
+    pub stability: f64,
+    pub difficulty: f64,
     pub elapsed_days: u32,
     pub scheduled_days: u32,
     pub reps: u32,
